Deploy VNPAY

+++ database/migrations/2024_08_02_021547_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id(); // ID
             $table->string('name'); // Name
-            $table->decimal('price', 8, 2); // Price
+            $table->integer('price'); // Price
             $table->text('description'); // Description
             $table->string('smell'); // Smell
             $table->string('texture'); // Texture
